requirements.txt
boto3
mail-parser
pydantic
python-dotenv
supabase
pypinyin
requests
fastapi
uvicorn
playwright
beautifulsoup4

